pyproject.toml
[project]
name = "backend-assistente"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "docling>=2.63.0",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.122.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-openai>=0.3.0",
    "langchain-huggingface>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-postgres>=3.0.1",
    "markdown>=3.10",
    "psycopg-binary>=3.2.13",
    "psycopg-pool>=3.2.8",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "tempfile2>=0.1.2",
    "uuid>=1.30",
    "uv>=0.9.13",
    "uvicorn>=0.38.0",
    "azure-storage-blob>=12.28.0",
]
